avs2yuv.c: replace ftime() with clock_gettime()
ftime() has been deprecated since POSIX.1-2001 and was removed in
POSIX.1-2008. While some platforms have kept compatibility with it
in the time since, the BSDs have removed it entirely, and glibc
deprecated it in v2.31.

An OLD_TIME_BEHAVIOR define has been added which forces the use of
the older ftime-based tactic; this required adding the ability for
user-passable CFLAGS to the Makefile.

OLD_TIME_BEHAVIOR exists as a define mostly to allow avs2yuv to be
used on versions of OS X prior to 10.12 Sierra.  Virtually all of
the other OSes (and newer versions of macOS) already have
clock_gettime.